optimize log

+++ src/event_trace/stack_walk.rs
@@ -24,7 +24,7 @@ impl<T: Clone> StackWalkMap<T> {
     pub fn insert(&mut self, key: (u32, i64), value: T, debug_msg: String) -> Option<(T, String)> {
         let old = self.events_map.insert(key, (value, debug_msg.clone()));
         if old.is_some() {
-            error!("Event's key for stack walk is repeated. thread_id: {} timestamp: {} {debug_msg}", key.0, key.1);
+            error!("Event's key for stack walk is repeated. thread_id: {} timestamp: {} {debug_msg}", key.0 as i32, key.1);
         }
         // clear events_map's item that is hold too long. Avoid map being too large
         self.pop_front(false, key.1);
